Rename add todo form component
